extract the notifier (sender) from the notifications (the content)
