+++ lib/ontologies_linked_data.rb
@@ -17,6 +17,7 @@ require "ontologies_linked_data/serializers/serializers"
 require "ontologies_linked_data/utils/file"
 require "ontologies_linked_data/utils/triples"
 require "ontologies_linked_data/utils/notifications"
+require "ontologies_linked_data/utils/notifier"
 require "ontologies_linked_data/utils/ontology_csv_writer"
 require "ontologies_linked_data/utils/multi_logger"
 require "ontologies_linked_data/parser/parser"
